broke out various endpoints to corresponding folders, added event task endpoints, added endpoint to mark a task as completed/incompelte
