+++ server/interfaces/request.interface.ts
@@ -3,7 +3,7 @@ import { iPermission } from "./permissions.interface";
 declare global {
   namespace Express {
     interface Request {
-      userId: string | undefined;
+      userId: number | undefined;
       permissions: iPermission[];
     }
   }
